Выполнил 3 задание
