+++ src/get-task-data.js
@@ -1,0 +1,31 @@
+export default () => ({
+  title: [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`][Math.floor(Math.random() * 3)],
+  dueDate: Date.now() + 1 + Math.floor(Math.random() * 7) * 24 * 60 * 60 * 1000,
+  tags: new Set([
+    `homework`,
+    `theory`,
+    `practice`,
+    `intensive`,
+    `keks`
+  ]),
+  picture: `//picsum.photos/100/100?r=${Math.random()}`,
+  color: new Set([
+    `black`,
+    `yellow`,
+    `blue`,
+    `green`,
+    `pink`
+  ]),
+  repeatingDays: {
+    'isRepeat': false,
+    'mo': true,
+    'tu': false,
+    'we': true,
+    'th': false,
+    'fr': false,
+    'sa': true,
+    'su': false,
+  },
+  isFavorite: true,
+  isDone: false
+});
